split the flow controller in stream and connection flow controller
